Refactor torch::jit::script::Module::register_* API.
Conceptually, Module is just a view over ClassType and ivalue::object.
register_ methods are the only methods that are exception from this:
they provide an API not available on ClassType or object directly. This
PR ports this API to ClassType and makes Module truly just a view over
those two.

ghstack-source-id: d1bf37a2c738af096ea1762c05d1668519385ac2
Pull Request resolved: https://github.com/pytorch/pytorch/pull/27189